move config to pkg
